backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
cryptography==43.0.1
python-dotenv==1.0.1
websockets==13.1
